requirements.txt
gcsa>=2.6.0
pyyaml>=6.0
python-dateutil>=2.8
